requirements.txt
ipywidgets==8.0.7
jupyter==1.0.0
langchain==0.1.8
langchai_ibm
matplotlib==3.7.2
numpy==1.24.2
pandas==1.5.3
plotly==5.15.0
pypdf==3.12.2
python-dotenv==1.0.0
pydantic==1.10.15
requests==2.31.0
urllib3==1.26.11
rouge==1.0.1
-U pymilvus==2.4.8
PyMuPDF
milvus==2.3.3
streamlit==1.35.0
huggingface-hub==0.23.2
sentence_transformers==2.7.0
-U langchain-community